Turn rest api into some basic web framework
